requirements.txt
django-environ==0.8.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pyserial==3.5
